Update v. 0.11.1: added 'highlighted comments'
Added highlighted comments: the route /blog-post/:id can now receive two additional query parameters: rootCommentAtTopId and highlightedCommentId. The first one is id of comment which is to be displayed above all other comments and the second one is id of comment which is to be highlighted (it may be above mentioned comment or one of its replies);
Improved notifications: when notification about new comment reply or new comment like is clicked, the above mentioned highlighted comment feature is used. The comment which has been liked (or a reply) is highlighted and scrolled to.